# public/swagger.json
{
    "openapi": "3.0.0",
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Registration"
                ],
                "summary": "Create a user",
                "description": "Returns Created User data",
                "operationId": "registerUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "King Salu"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "kingsalu@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "King Salu"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "kingsalu@example.com"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "datetime",
                                            "example": "2025-01-02T21:47:26.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "datetime",
                                            "example": "2025-01-02T21:47:26.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/validate": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Token validation.",
                "description": "Returns validation and User data",
                "operationId": "validateToken",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Token is valid"
                                        },
                                        "user": {
                                            "type": "object",
                                            "example": {
                                                "id": 1,
                                                "name": "King Salu",
                                                "email": "kingsalu@exacmple.com",
                                                "email_verified_at": null,
                                                "created_at": "2025-01-01T18:05:07.000000Z",
                                                "updated_at": "2025-01-01T18:05:07.000000Z"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Token not provided"
                    },
                    "401": {
                        "description": "Invalid user"
                    }
                }
            }
        },
        "/api/order/initiate": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Start an order if balance permits.",
                "description": "Returns Created Order data",
                "operationId": "initiateOrder",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "item",
                                    "price"
                                ],
                                "properties": {
                                    "item": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Item 1"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "3000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "user_id": 1,
                                                "item": "item x",
                                                "price": 1000,
                                                "status": "success",
                                                "updated_at": "2025-01-03T08:10:47.000000Z",
                                                "created_at": "2025-01-03T08:10:47.000000Z"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Insufficient Funds"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/order/status": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Retrieve order status.",
                "description": "Returns latest Order data",
                "operationId": "orderStatus",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user_id": {
                                            "type": "integr",
                                            "example": "1"
                                        },
                                        "item": {
                                            "type": "string",
                                            "example": "item x"
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": "1000"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "pending"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2025-01-03T08:10:47.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2025-01-03T08:10:47.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/wallet/balance": {
            "get": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Fetch user balance",
                "description": "Returns wallet information data from logged in user",
                "operationId": "walletBalance",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "status": {
                                            "type": "boolean",
                                            "example": "1"
                                        },
                                        "balance": {
                                            "type": "number",
                                            "format": "float",
                                            "example": "20000"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2025-01-02T21:47:26.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2025-01-02T21:47:26.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/wallet/transfer": {
            "post": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Simulate wallet transfer",
                "description": "Returns message showing transfer successful",
                "operationId": "transfer",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "to"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "10000"
                                    },
                                    "to": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "kingsalu@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Transfer Successful"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Insufficient Funds"
                    },
                    "404": {
                        "description": "Beneficiary Not Found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login user",
                "description": "Returns wuthentication data",
                "operationId": "loginUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "access_token": {
                                            "type": "string",
                                            "example": "{AlphaNumeric code}"
                                        },
                                        "refresh_token": {
                                            "type": "string",
                                            "example": "{AlphaNumeric code}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "The user credentials were incorrect."
                    }
                }
            }
        },
        "/api/refresh": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh",
                "description": "Returns 'Refreshed'",
                "operationId": "refreshToken",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "refresh_token"
                                ],
                                "properties": {
                                    "refresh_token": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "{Alphanumeric code}"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string",
                                    "example": "Refreshed"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User not authenticated"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Registration",
            "description": "Registration"
        },
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "Order",
            "description": "Order"
        },
        {
            "name": "Wallet",
            "description": "Wallet"
        }
    ]
}